SDK-604: Move sandbox out of src
