+++ sandbox/profile/request/attribute/derivation/age.verification.js
@@ -1,8 +1,8 @@
 const SandboxAttributeBuilder = require('../attribute.builder');
 const SandboxAnchor = require('../anchor');
-const { YotiDate } = require('../../../../../data_type/date');
-const Validation = require('../../../../../yoti_common/validation');
-const constants = require('../../../../../yoti_common/constants');
+const { YotiDate } = require('../../../../../src/data_type/date');
+const Validation = require('../../../../../src/yoti_common/validation');
+const constants = require('../../../../../src/yoti_common/constants');
 
 /**
  * @class SandboxAgeVerification
